# backend/openapi.yaml
openapi: 3.0.0
info:
  title: BrainFood API
  description: |
    REST API für die BrainFood Vokabel-App mit FSRS-5 Spaced Repetition Algorithmus.
    
    Diese API ermöglicht die Verwaltung von Boxen, Karteikarten und die intelligente
    Wiederholungsplanung basierend auf dem FSRS-5 Algorithmus.
  version: 1.0.0
  contact:
    name: BrainFood API Support
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-domain.com/api
    description: Production server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    Box:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    Card:
      type: object
      properties:
        id:
          type: string
        boxId:
          type: string
        front:
          type: string
          description: Question or word to learn
        back:
          type: string
          description: Answer or translation
        tags:
          type: string
          nullable: true
          description: Comma-separated tags
        stability:
          type: number
          description: FSRS-5 stability value
        difficulty:
          type: number
          description: FSRS-5 difficulty value (0-1)
        reps:
          type: integer
          description: Number of successful repetitions
        lapses:
          type: integer
          description: Number of failures
        lastReviewAt:
          type: string
          format: date-time
          nullable: true
        due:
          type: string
          format: date-time
          description: Next due date for review
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT access token
    CreateBoxRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    CreateCardRequest:
      type: object
      required:
        - front
        - back
      properties:
        front:
          type: string
        back:
          type: string
        tags:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account and returns a JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticates a user and returns a JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Returns information about the currently authenticated user
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /boxes:
    get:
      tags:
        - Boxes
      summary: Get all boxes
      description: Returns all boxes for the current user
      responses:
        '200':
          description: List of boxes
          content:
            application/json:
              schema:
                type: object
                properties:
                  boxes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Box'

    post:
      tags:
        - Boxes
      summary: Create a new box
      description: Creates a new box for organizing flashcards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoxRequest'
      responses:
        '201':
          description: Box created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  box:
                    $ref: '#/components/schemas/Box'
  /boxes/{boxId}:
    put:
      tags:
        - Boxes
      summary: Update a box
      description: Updates the name of a box
      parameters:
        - name: boxId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoxRequest'
      responses:
        '200':
          description: Box updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  box:
                    $ref: '#/components/schemas/Box'

    delete:
      tags:
        - Boxes
      summary: Delete a box
      description: Deletes a box and all its cards
      parameters:
        - name: boxId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Box deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /boxes/{boxId}/cards:
    get:
      tags:
        - Cards
      summary: Get all cards in a box
      description: Returns all cards in a specific box, optionally filtered by search query
      parameters:
        - name: boxId
          in: path
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
          description: Search term to filter cards
        - name: sort
          in: query
          schema:
            type: string
            enum: [due]
          description: Sort order (by due date)
      responses:
        '200':
          description: List of cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'

    post:
      tags:
        - Cards
      summary: Create a new card
      description: Creates a new flashcard in a box with initial FSRS-5 state
      parameters:
        - name: boxId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  card:
                    $ref: '#/components/schemas/Card'
  /cards/{cardId}:
    get:
      tags:
        - Cards
      summary: Get card details
      description: Returns detailed information about a specific card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Card details
          content:
            application/json:
              schema:
                type: object
                properties:
                  card:
                    $ref: '#/components/schemas/Card'

    put:
      tags:
        - Cards
      summary: Update a card
      description: Updates the front, back, or tags of a card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '200':
          description: Card updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  card:
                    $ref: '#/components/schemas/Card'

    delete:
      tags:
        - Cards
      summary: Delete a card
      description: Deletes a flashcard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Card deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...